Create Dockerfile and build script to container-ize flowlib

+++ flowlib/new/util.py
@@ -3,9 +3,14 @@ import subprocess
 import sys
 import json
 
-def call_cmd(container, endpoint, cmd):
-    cmd = "docker run -it --rm --entrypoint /opt/nifi/nifi-toolkit-current/bin/cli.sh {container} {cmd} --baseUrl {endpt} -ot json"\
-        .format(container=container, cmd=cmd, endpt=endpoint)
+def call_cmd(container, endpoint, command):
+    cmd = None
+    if container:
+        cmd = "docker run -it --rm --entrypoint /opt/nifi/nifi-toolkit-current/bin/cli.sh {container} {cmd} --baseUrl {endpt} -ot json"\
+            .format(container=container, cmd=command, endpt=endpoint)
+    else:
+        cmd = "/opt/nifi/nifi-toolkit-current/bin/cli.sh {cmd} --baseUrl {endpt} -ot json" \
+            .format(container=container, cmd=command, endpt=endpoint)
     cmd_output = subprocess.Popen(cmd, shell=True, stdout=subprocess.PIPE, stderr=subprocess.STDOUT, encoding='UTF-8')
     response = cmd_output.communicate()[0]
     if response.strip().startswith("ERROR:"):
